Now some errors are displayed, check now if the .json file is correctly written. Added new program to convert and updated guide.
Now errors are displayed in case a pdf file does not exist for conversion, if the page does not exist (so if a document has 5 pages and you insert a sixth page), if the file at the time of renaming already exists one with an equal number.
Added checking of the .json file, to make sure there are no write errors.
Added program to convert a pdf test file to jpg to take all the points to put in the .json file.
Updated the README, read it to get more information about the .json file.